SoundMorph Example
This is an example of how a sound morph can be done using the sounds used for the final project, it makes use of the Pd library "FFTease" which has been included as a weblink in the .txt file in this folder. In the morphing.pd patch, after the two sounds are initialised a slider can be slid to morph from one sound to the other, moving their peaks.